+++ src/main/java/com/mamiyaotaru/voxelmap/gui/GuiSlotWaypoints.java
@@ -181,7 +181,7 @@ class GuiSlotWaypoints extends GuiSlotMinimap {
             int leftEdge = this.parentGui.getWidth() / 2 - 92 - 16;
             byte padding = 3;
             int width = 215;
-            if (mouseX >= (double) (leftEdge + width - 16 - padding) && mouseX <= (double) (leftEdge + width + padding)) {
+            if (mouseX >= (leftEdge + width - 16 - padding) && mouseX <= (leftEdge + width + padding)) {
                 if (GuiSlotWaypoints.this.doubleclick) {
                     this.parentGui.setHighlightedWaypoint();
                 }
